requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
streamlit==1.44.1
requests==2.32.3
python-multipart==0.0.20
python-dotenv==1.0.1
langchain-core==0.3.56
langchain-community==0.3.21
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
faiss-cpu==1.10.0
openai==1.75.0
pypdf==5.4.0
python-docx==1.1.2
pandas==2.2.3
python-pptx==1.0.2
    